+++ src/KokkosComm/mpi/alltoall.hpp
@@ -36,10 +36,10 @@
 
 #include <Kokkos_Core.hpp>
 
-#include "KokkosComm_traits.hpp"
-#include "impl/KokkosComm_pack_traits.hpp"
-#include "impl/KokkosComm_include_mpi.hpp"
-#include "impl/KokkosComm_types.hpp"
+#include "KokkosComm/traits.hpp"
+#include "impl/pack_traits.hpp"
+#include "impl/include_mpi.hpp"
+#include "impl/types.hpp"
 
 namespace KokkosComm::Impl {
 template <KokkosExecutionSpace ExecSpace, KokkosView SendView, KokkosView RecvView>
