+++ src/KokkosComm/mpi/barrier.hpp
@@ -16,19 +16,26 @@
 
 #pragma once
 
-#include <utility>
-
-#include <Kokkos_Core.hpp>
-
-#include "KokkosComm_fwd.hpp"
-#include "KokkosComm_concepts.hpp"
+#include "KokkosComm/concepts.hpp"
 
 namespace KokkosComm {
 
-template <KokkosExecutionSpace ExecSpace = Kokkos::DefaultExecutionSpace,
-          CommunicationSpace CommSpace   = DefaultCommunicationSpace>
-void barrier(Handle<ExecSpace, CommSpace> &&h) {
-  Impl::Barrier<ExecSpace, CommSpace>{std::forward<Handle<ExecSpace, CommSpace>>(h)};
+namespace Impl {
+template <KokkosExecutionSpace ExecSpace, CommunicationSpace CommSpace>
+struct Barrier {
+  Barrier(Handle<ExecSpace, Mpi> &&h) {
+    h.space().fence("KokkosComm::Impl::Barrier");
+    MPI_Barrier(h.mpi_comm());
+  }
+};
+}  // namespace Impl
+
+namespace mpi {
+inline void barrier(MPI_Comm comm) {
+  Kokkos::Tools::pushRegion("KokkosComm::mpi::barrier");
+  MPI_Barrier(comm);
+  Kokkos::Tools::popRegion();
 }
+}  // namespace mpi
 
 }  // namespace KokkosComm
