+++ unit_tests/test_sendrecv.cpp
@@ -16,7 +16,7 @@
 
 #include <gtest/gtest.h>
 
-#include "KokkosComm.hpp"
+#include "KokkosComm/KokkosComm.hpp"
 
 #include "view_builder.hpp"
